requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
PyMuPDF
pytesseract
numpy
scikit-learn
Pillow